Move level up cost responsibility to Champions module
