+++ apps/champions/lib/champions/utils.ex
@@ -1,0 +1,5 @@
+defmodule Champions.Utils do
+  @moduledoc false
+
+  def game_id(), do: 2
+end
