+++ apps/game_backend/lib/game_backend/transaction.ex
@@ -1,0 +1,11 @@
+defmodule GameBackend.Transaction do
+  @moduledoc """
+  Module for exponsing transactions to game applications.
+  """
+
+  alias GameBackend.Repo
+
+  def run(transaction) do
+    Repo.transaction(transaction)
+  end
+end
